+++ projects/ngx-form-object-sample-app/src/app/models/user.form-object.ts
@@ -4,12 +4,16 @@ import { FormObject, FormObjectOptions } from 'ngx-form-object';
 import { Observable, of } from 'rxjs';
 import { User } from './user.model';
 
-export class UserFormObject extends FormObject {
+export class UserFormObject extends FormObject<User> {
 	public validators: Record<string, ValidatorFn | Array<ValidatorFn>> = {
 		name: [Validators.required],
 	};
 
-	constructor(public model: any, protected options: FormObjectOptions, public injector: Injector) {
+	constructor(
+		public model: User,
+		protected options: FormObjectOptions,
+		public injector: Injector
+	) {
 		super(model, options);
 	}