+++ projects/ngx-form-object/src/lib/form-store/form-store.ts
@@ -2,23 +2,23 @@ import { FormGroup } from '@angular/forms';
 import { Observable } from 'rxjs';
 import { FormObject } from '../form-object/form-object';
 
-export class FormStore extends FormGroup {
-	private _formObject: FormObject;
-	private _isSubmitted = false;
+export class FormStore<T> extends FormGroup {
+	private _formObject: FormObject<T>;
+	private _isSubmitted: boolean = false;
 
 	public get isChanged(): boolean {
 		return this.attributesDidChange || this.belongsToPropertiesDidChange || this.hasManyPropertiesDidChange;
 	}
 
-	public set formObject(formObject: FormObject) {
+	set formObject(formObject: FormObject<T>) {
 		this._formObject = formObject;
 	}
 
-	public get formObject(): FormObject {
+	get formObject(): FormObject<T> {
 		return this._formObject;
 	}
 
-	public get model(): any {
+	get model(): T {
 		return this.formObject.model;
 	}
 
@@ -26,7 +26,7 @@ export class FormStore extends FormGroup {
 		return this._isSubmitted;
 	}
 
-	public save(): Observable<any> {
+	public save(): Observable<T> {
 		this._isSubmitted = true;
 		return this.formObject._save(this);
 	}