+++ CHANGELOG.md
@@ -58,6 +58,11 @@
 
   * The type of validators is adjusted
 
+### 8.0.0-beta
+
+  * [BREAKING CHANGE] `FormStore` uses a generic now
+  * [BREAKING CHANGE] `FormObject` uses a generic now
+
 ### 9.0.3-beta
 
   * Temporary adjust save method types